Renamed yield policies, optimized string pool
Yield policies have been renamed, old names are deprecated.
StringPool now uses a hash rather than a brute-force search.
Fixed some incorrect uses of #elif defined() on theron defines, which
are always defined.

+++ Include/Theron/Detail/Transport/InputMessage.h
@@ -98,7 +98,7 @@ public:
 
 #if THERON_XS
 
-        size = xs_msg_size(&mMessage);
+        size = static_cast<uint32_t>(xs_msg_size(&mMessage));
 
 #endif // THERON_XS
